+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ are used for versioning (schema follows below):
 - Make it possible to fetch additional information (for now from Jira only,
   but can be extended).
 - Softened the regular expression patterns for ticket numbers/branch names.
+- Initial steps on on introducing renderer classes.
 
 0.3.4
 -----