+++ CHANGELOG.rst
@@ -15,8 +15,8 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
-0.3.5
-------
+0.4
+---
 2019-12-20
 
 - Placing ``.matyan.ini`` config file (placed in the home root directory